Move showName to scala.internal
