+++ docs/docs/reference/metaprogramming/macros.md
@@ -564,53 +564,6 @@ while (i < arr.length) {
 }
 sum
 ```
-### Showing meaningful definition names in quotes
-
-In the `powerCode` example above there is a `'{ val y = $x * $x; ... }` which when printed
-may show several different `val y = ...`.
-
-For example
-```scala
-powerCode(16, '{7}).show
-```
-will show
-```scala
-val y: scala.Double = 7
-val y: scala.Double = y.*(y)
-val y: scala.Double = y.*(y)
-val y: scala.Double = y.*(y)
-val y: scala.Double = y.*(y)
-y
-```
-Even though there is no hygiene issue it may be hard to undestand the code. To overcome this inconvenience
-each `y` can be assigned a meaningful name using the `scala.quoted.show.showName` annotation.
-For example `'{ @showName(${Expr("y" + i)}) val y = $x * $x; ... }` will assign to each `y` a name
-`a{i}` where `{i}` is a known String, if `i == 3` then it would be named `a3`.
-
-The `powerCode` can be defined as follows using `showName`
-```scala
-def powerCodeD(n: Long, x: Expr[Double]))(given QuoteContext): Expr[Double] = '{
-  val a1 = $x
-  ${ powerCodeD(n, 2, 'x1) }
-}
-def powerCodeD(n: Long, i: Int, x: Expr[Double])(given QuoteContext): Expr[Double] =
-  if (n == 0) '{1.0}
-  else if (n % 2 == 0) '{ @showName(${Expr("a" + i)}) val y = $x * $x; ${powerCodeD(n / 2, idx * 2, 'y)} }
-  else '{ $x * ${powerCodeD(n - 1, idx, x)} }
-```
-then
-```scala
-powerCodeD(16, '{7}).show
-```
-will show
-```scala
-val a1: scala.Double = 7
-val a2: scala.Double = a1.*(a1)
-val a4: scala.Double = a2.*(a2)
-val a8: scala.Double = a4.*(a4)
-val a16: scala.Double = a8.*(a8)
-a16
-```
 
 ### Find implicits within a macro
